www/requirements.txt
mkdocs==1.6.1
mkdocs-material==9.6.16
mkdocs-material-extensions==1.3.1
Pillow==12.1.1
